implemented reconfigurer
