+++ service/reconfigurer/reconfigurer.go
@@ -7,5 +7,5 @@ import (
 // Provider describes a type that can provide config state events to a target
 // watcher. It will reconfigure and restart the watcher whenever necessary.
 type Provider interface {
-	Configure(watcher.Watcher)
+	Configure(watcher.Watcher) error
 }
